Split  spawnpoint into smaller parts
