+++ lib/spawnpoint.js
@@ -1,0 +1,980 @@
+'use strict';
+
+// Include core libs
+const fs = require('fs'),
+	child_process = require('child_process'),
+	crypto = require('crypto'),
+	util = require('util'),
+	path = require('path'),
+	assert = require('assert'),
+	EventEmitter = require('events');
+
+// Include external libraries
+const _ = require('lodash'),
+	async = require('async'),
+	minimist = require('minimist'),
+	moment = require('moment'),
+	containerized = require('containerized');
+
+// Define private helper functions
+const helpers = require('./helpers.js');
+
+/**
+ * Agnostic JS framework that empowers devs to focus on quickly building apps, rather than focusing on application
+ * config, health-checks, application structure, or architecture to build a 12 factor app in Docker.
+ *
+ * Spawnpoint can be configured to manage the entire application life-cycle or standalone as a utility library.
+ * @class
+ */
+class spawnpoint extends EventEmitter{
+	/**
+	 * Creates new instance of spawnpoint
+	 * @param  {string} [configFile] Sets the JSON file spawnpoint uses to setup the framework.
+	 * @return {this}
+	 */
+	constructor(configFile = '/config/app.json'){
+		// init EventEmitter
+		super();
+
+		// set the spawnpoint config file
+		this.configFile = configFile;
+
+		// Used to track if the application expects itself to continue running or not
+		this.status = {
+			setup: false,
+			running: false,
+			stopping: false,
+			stopAttempts: 0
+		};
+		this.cwd = process.cwd();
+
+		// detect if we are in a container
+		this.containerized = containerized();
+
+		// list of ENV configs that are blacklisted
+		this.configBlacklist = require('../config-blacklist.json');
+
+		// plugin registery
+		this.register = [];
+
+		// config object to store all application config
+		this.config = {};
+
+		// codes object to store all Spawnpoint codes
+		this.codes = {};
+
+		// errorMaps help wrap custom Error types to Spawnpoint codes.
+		this.errorMaps = {};
+
+		// error tracking, debounce detection
+		this.limitMaps = {};
+
+		// which plugins are loaded
+		this.plugins = {};
+
+
+		// make errorCode and failCode available
+		this._errorCode = require('./errorCode.js');
+		this._failCode = require('./failCode.js');
+
+		// log formatting
+		this.logs = {
+			prefix: null,
+			date: null
+		};
+
+		return this;
+	}
+
+	/**
+	 * Initializes framework to read the `configFile`, init config, Spawnpoint plugins, errorCodes and autoload
+	 * folders. This also starts the application life-cycle so the app can stop gracefully.
+	 * @callback {Function} [callback] Triggered once the `app.ready` event triggers.
+	 * @return {this}
+	 */
+	setup(callback = () => {}){
+		var self = this;
+
+		// force .json parsing with comments :)
+		self.setupJSONHandler();
+
+		// prevent repeated setup
+		if(self.status.setup){
+			return callback(self.code('app.already_setup'));
+		}
+		self.status.setup = true;
+
+		// App loading process
+		self.initConfig();
+		self.initCodes();
+		self.initRegistry();
+		self.loadPlugins();
+		self.loadConfig();
+		self.loadCodes();
+		self.initLimitListeners();
+		self.loadErrorMap();
+		var jobs = [];
+
+		_.each(self.plugins, function(plugin){
+			if(plugin.callback){
+				return jobs.push(function(cb){
+					return plugin.exports(self, cb);
+				});
+			}
+			jobs.push(function(cb){
+				plugin.exports(self);
+				return cb();
+			});
+		});
+		// load framework files
+		_.each(self.config.autoload, function(jobDetails){
+			self.log('Autoloading %s', jobDetails.name || jobDetails.folder);
+			var list = self.recursiveList(util.format('%s/%s', self.cwd, jobDetails.folder), jobDetails.extension || '.js');
+			if(jobDetails.callback){
+				return jobs.push(function(callback){
+					async.eachSeries(list, function(file, acb){
+						var error;
+						try{
+							require(file)(self, acb);
+						}catch(err){
+							error = err;
+						}
+						if(error){
+							return acb(error);
+						}
+					}, callback);
+				});
+			}
+			jobs.push(function(callback){
+				_.each(list, function(file){
+					var error;
+					try{
+						require(file)(self);
+					}catch(err){
+						error = err;
+					}
+					if(error){
+						return console.error(error);
+					}
+				});
+				return callback();
+			});
+		});
+		process.nextTick(function(){
+			async.series(jobs, function(err){
+				if(err){
+					self.error("Failed to start up").debug(err);
+					self.emit('app.exit');
+					return callback(err);
+				}
+				self.log('%s is ready.', self.config.name);
+				self.emit('app.ready');
+				return callback();
+			});
+		});
+		self.emit('app.setup.done');
+		return this;
+	}
+
+	/**
+	 * Recursively list files in a directory by an optional file extension.
+	 * NOTE: This is an event blocking sync method.
+	 * @param  {String} dir  Directory to list files from.
+	 * @param  {Array|string} [exts] Optional list of file extensions to return. Defaults to .js files. Set to a falsy value to disable this filter.
+	 * @return {Array} Absolute/full path of filenames found.
+	 */
+	recursiveList(dir, exts = ['.js']){
+		if(typeof(exts) === 'string'){
+			exts = [exts];
+		}
+		var parent = this, stat, list = [];
+		try{
+			stat = fs.statSync(dir);
+		}catch(e){
+			stat = false;
+		}
+		if(!stat || !stat.isDirectory()){
+			return list;
+		}
+		dir = String(dir + '/').replace(/\//g, '/'); // ensure proper trailing slash
+		_.each(fs.readdirSync(dir), function(file){
+			var isDir = fs.statSync(dir + file).isDirectory();
+			if(isDir && exts && exts.indexOf('/') !== -1){
+				list.push(dir + file);
+			}else if(isDir){
+				var recursive = parent.recursiveList(dir + file, exts);
+				if(recursive instanceof Array && recursive.length > 0){
+					list = list.concat(recursive); // add results
+				}
+			}else{
+				if(!exts || exts.indexOf(path.extname(file)) !== -1){
+					list.push(dir + file);
+				}
+			}
+		});
+		list.sort(); // windows won't sort this like unix will
+		return list;
+	}
+
+	/**
+	 * Utility: Create random string.
+	 * @param  {Number} [length] How long of a random string to create.
+	 * @param  {String} [hashMethod] Which crypto hash method to use.
+	 * @return {String} Random string of characters.
+	 */
+	random(length = 16, hashMethod = 'md5'){
+		length = parseInt(length);
+		if(isNaN(length) || length < 1){
+			length = 16; // TODO: throw an error in an update
+		}
+		var random = String(new Date().getTime() + Math.random());
+		return String(crypto.createHash(hashMethod).update(random).digest('hex')).slice(0, length);
+	}
+
+	/**
+	 * Utility: get random element from `collection`.
+	 * This is a copy of the lodash _.sample method.
+	 * @param  {Array|Object} items The collection to sample.
+	 * @return {*} Returns the random element.
+	 */
+	sample(items){
+		return _.sample(items);
+	}
+
+	/**
+	 * Utility: Creates new `roundRobin` class with collection.
+	 * @param  {Array|Object} items The collection to sample.
+	 * @return {roundRobin} Returns new instance of `roundRobin` class.
+	 */
+	roundRobbin(items){
+		return new require('./roundRobbin.js')(this)(items);
+	}
+
+	/**
+	 * Utility: get random element from `collection` in an async lock.
+	 * @param  {Array|Object} items The collection to sample.
+	 * @return {roundRobin} Returns new instance of `roundRobin` class.
+	 */
+	getAndLock(items){
+		return new require('./getAndLock.js')(this)(items);
+	}
+
+	/**
+	 * Checks if the current application runtime is running as a root user/group.
+	 * @return {Boolean} When true: the application is running as a root user/group.
+	 */
+	isRoot(){
+		if(this.isSecure() === true){
+			return false;
+		}else{
+			return true;
+		}
+	}
+
+	/**
+	 * Checks if the current application runtime is running as a specific `uid` and/or `gid`.
+	 * @param  {Number}  [uid] Unix `uid` to check against.
+	 * @param  {Number}  [gid] Unix `gid` to check against. When not set will match `uid`.
+	 * @return {Boolean} When true: the application is running as the user/group.
+	 */
+	isSecure(uid, gid){
+		var self = this;
+		if(uid && !gid){
+			gid = uid;
+		}
+		var checks = {
+			uid: process.getuid(),
+			gid: process.getgid(),
+			groups: String(child_process.execSync('groups'))
+		};
+		if(checks.uid === 0 || checks.gid === 0){
+			return self.errorCode('usercheck.is_root', {checks: checks });
+		}
+		if(checks.groups.indexOf('root') !== -1){
+			return self.errorCode('usercheck.is_root_group', {checks: checks });
+		}
+		if(uid && gid && (uid !== checks.uid || gid !== checks.gid)){
+			return self.errorCode('usercheck.incorrect_user', {checks: checks });
+		}
+		return true;
+	}
+
+	/**
+	 * Helper method that requires a file and hoists the current spawnpoint application reference.
+	 * @param  {String} path File path to require.
+	 */
+	require(path){
+		var self = this;
+		return require(path)(self);
+	}
+
+	/**
+	 * Builds a Spawnpoint code object. Codes are used to create a link between a human readable message
+	 * and a computer readable string. Example: `file.not_found` -> "The requested file was not found."
+	 * @param {String} code computer readable string code.
+	 * @param {Object} [data] Object to extend the code Object with
+	 * @return {Object} Code Object with a `message` with the computer readable message and the `code` matching the input code.
+	 */
+	code(code, data = {}){
+		if(!this.codes[code]){
+			throw new Error("No return code found with code: " + code); // TODO: convert this to an errorCode
+		}
+		return _.defaults(data, {
+			code: code,
+			message: this.codes[code]
+		});
+	}
+
+	/**
+	 * Spawnpoint code that wraps a Javascript `Error` as a hard application error.
+	 * @param {String} code computer readable string code.
+	 * @param {Object} [data] Object to extend the code Object with
+	 * @return {Object} Error Code Object with a `message` with the computer readable message and the `code` matching the input code.
+	 */
+	errorCode(code, data){
+		var getCode = this.code(code, data);
+		this.emit('errorCode', getCode);
+		return new this._errorCode(getCode);
+	}
+
+	/**
+	 * Spawnpoint code that wraps a Javascript `Error`, as a soft error.
+	 * @param {String} code computer readable string code.
+	 * @param {Object} [data] Object to extend the code Object with
+	 * @return {Object} Error Code Object with a `message` with the computer readable message and the `code` matching the input code.
+	 */
+	failCode(code, data){
+		var getCode = this.code(code, data);
+		this.emit('failCode', getCode);
+		return new this._failCode(getCode);
+	}
+
+	/**
+	 * Console.log wrapper that only triggers with when `config.debug` is enabled.
+	 * @params {*} [args..] Arguments to be passed to logging.
+	 * @return {this}
+	 */
+	debug(){
+		if(this.config.debug){
+			console.log.apply(this, arguments);
+		}
+		return this;
+	}
+
+	/**
+	 * Console.log wrapper that adds an INFO tag and timestamp to the log.
+	 * @params {String|Object|Array|Number} [args..] Arguments to be passed to logging.
+	 * @return {this}
+	 */
+	info(){
+		helpers.log({
+			logs: this.logs,
+			config: this.config.log,
+			type: helpers.tag('INFO', chalk.green),
+			line: chalk.white(util.format.apply(this, arguments))
+		});
+		return this;
+	}
+
+	/**
+	 * Console.log wrapper that adds an LOG tag and timestamp to the log.
+	 * @params {String|Object|Array|Number} [args..] Arguments to be passed to logging.
+	 * @return {this}
+	 */
+	log(){
+		helpers.log({
+			logs: this.logs,
+			config: this.config.log,
+			type: helpers.tag('LOG', chalk.cyan),
+			line: chalk.white(util.format.apply(this, arguments))
+		});
+		return this;
+	}
+
+	/**
+	 * Console.error` wrapper that adds an WARN tag and timestamp to the log. This prints to STDERR.
+	 * @params {String|Object|Array|Number} [args..] Arguments to be passed to logging.
+	 * @return {this}
+	 */
+	warn(){
+		helpers.log({
+			logs: this.logs,
+			config: this.config.log,
+			type: helpers.tag('WARN', chalk.yellow),
+			line: chalk.yellow(util.format.apply(this, arguments))
+		});
+		return this;
+	}
+
+	/**
+	 * Console.error` wrapper that adds an ERROR tag and timestamp to the log. This prints to STDERR.
+	 * @params {String|Object|Array|Number} [args..] Arguments to be passed to logging.
+	 * @return {this}
+	 */
+	error(){
+		helpers.log({
+			logs: this.logs,
+			config: this.config.log,
+			type: helpers.tag('ERROR', chalk.red.bold),
+			line: chalk.red(util.format.apply(this, arguments))
+		}, 'error');
+		return this;
+	}
+
+	/**
+	 * Registers multiple custom Errors to a specific errorCode. This helps wrap errors into a singular errorCode system.
+	 * @param {String} The errorCode human readable Spawnpoint code.
+	 * @param {Error} Instance of the error to map to..
+	 * @return {this}
+	 */
+	registerError(code, error){
+		this.errorMaps[code] = error;
+		return this;
+	}
+
+	/**
+	 * Registers multiple custom Errors to a specific errorCode, using the `registerError` method.
+	 * @param  {Object} errors Errors being registered. Each index/key is the errorCode string that the custom Error represents. The Value must be an uninitialized instance of the error.
+	 * @return {this}
+	 */
+	registerErrors(errors){
+		_.each(errors, (error, code) => {
+			this.registerError(code, error);
+		});
+		return this;
+	}
+
+	/**
+	 * Checks for Spawnpoint wrapped code, errorCode, or failCode when a potential error map is found (and previously registered). This method is useful as middleware to your application
+	 * error handling so that you don't have to have the server reply with a generic error.
+	 * @param  {Error} error Error to check for mapped error.
+	 * @return {errorCode|false} Returns Spawnpoint mapped code, errorCode, or failCode or false when no mapped error was found.
+	 */
+	maskErrorToCode(error, type = 'code'){
+		const validTypes = ['errorCode', 'failCode', 'code'];
+		let returnedError = false;
+		if(!validTypes.includes(type)){
+			throw new Error('Invalid `type` provided. Valid types:' + validTypes.join(',')); // TODO: convert to errorCode
+		}
+		_.each(self.errorMaps, (error, code) => {
+			if(!returnedError && error instanceof error){
+				returnedError = this[code](code, error);
+			}
+		});
+		return returnedError;
+	}
+
+	/**
+	 * Internal: Initializes the Spawnpoint `config` object. Reads package.json and `configFile` file to build config.
+	 * This is step 1 of 8 to startup Spawnpoint on line #157
+	 * @param  {String} [configFile]  Sets the JSON file Spawnpoint uses to setup the framework.
+	 * @private
+	 * @return {this}
+	 */
+	initConfig(configFile = null){
+		var self = this;
+		if(configFile){
+			this.configFile = configFile;
+		}
+		// reset config variable for reloading
+		this.config = _.defaults(require(this.cwd + this.configFile), {
+			debug: false,
+			plugins: [],
+			autoload: [],
+			secrets: '/run/secrets',
+			codes: '/config/codes',
+			configOverride: null,
+			stopAttempts: 3,
+			stopTimeout: 15000,
+			trackErrors: false,
+			log: {
+				format: '{date} {type}: {line}',
+				time: "HH:mm:ss",
+				date: "dddd, MMMM Do YYYY"
+			}
+		});
+		if(this.config.debug){
+			if(!this.config.configOverride){
+				this.config.configOverride = 'dev-config.json';
+			}
+		}
+		if(this.config.resetConfigBlackListDefaults){
+			this.configBlacklist = {
+				env: {list: [], patterns: []},
+				secrets: {list: [], patterns: []},
+				args: {list: [], patterns: []}
+			};
+		}
+		if(this.config.configBlacklist){
+			_.merge(this.configBlacklist, this.config.configBlacklist);
+		}
+		_.each(this.configBlacklist, function(items){
+			items.patterns = _.map(items.patterns, function(pattern){
+				return new RegExp(pattern);
+			});
+		});
+		var packageData = {};
+		try{
+			packageData = require(path.join(this.cwd, '/package.json'));
+		}catch(e){
+			// do nothing
+		}
+		// allow package.json version & name to set app.config vars
+		if(packageData.version){
+			this.config.version = this.config.version || packageData.version;
+		}
+		if(packageData.name){
+			this.config.name = this.config.name || packageData.name || "unnamed project";
+		}
+
+		// setup all of the required functions mounted on the `config` object
+
+		/**
+		 * Helper method to safely get a nested config item.
+		 * @param  {String} path The path of the property to get.
+		 * @param  {*} [defaultValue=false] The value returned for undefined resolved values.
+		 * @return {*} Returns the resolved value.
+		 */
+		this.config.get = function(path, defaultValue){
+			return _.get(self.config, path, defaultValue);
+		};
+
+		/**
+		 * Helper method to safely check if a nested config item exists
+		 * @param  {String} path The path to check.
+		 * @return {*} Returns `true` if path exists, else `false`.
+		 */
+		this.config.has = function(path){
+			return _.has(self.config, path);
+		};
+
+		/**
+		 * Helper method to get a random element from a Spawnpoint `config` item `collection`.
+		 * @param  {path} path The path to return items from.
+		 * @return {*} Returns random element from the collection.
+		 */
+		this.config.getRandom = function(path){
+			let items = self.config.get(path);
+			if(!items){
+				throw new self._errorCode('app.config.sample_not_array'); // TODO: choose better name
+			}
+			return _.sample(items);
+		};
+
+		let rrKeys = {};
+		/**
+		 * Helper method to get get random element from Spawnpoint `config` item `Array` with Round Robin ordering
+		 * This ensures no single item is returned more than it's siblings.
+		 * @param  {path} path The path to return items from.
+		 * @return {*} Returns random element from the collection.
+		 */
+		this.config.getRoundRobbin = function(path){
+			if(!rrKeys[path]){
+				let items = self.config.get(path);
+				if(!items || !items.length){
+					throw new self._errorCode('app.config.sample_not_array');
+				}
+				rrKeys[path] = self.roundRobbin(items);
+			}
+			return rrKeys[path].next();
+		};
+
+		let lockedKeys = {};
+		this.config.getAndLock = function(key, timeout, callback){
+			if(!lockedKeys[path]){
+				let items = self.config.get(path);
+				if(!items || !items.length){
+					throw new self._errorCode('app.config.sample_not_array');
+				}
+				lockedKeys[path] = self.getAndLock(items);
+			}
+			return lockedKeys[path].next(timeout, callback);
+		};
+
+		this.emit('app.setup.initConfig');
+		return this;
+	}
+
+	registerConfig(name, config, whiteListCheck){
+		var self = this,
+			data = {};
+
+		if(whiteListCheck && self.configBlacklist[whiteListCheck]){
+			if(self.configBlacklist[whiteListCheck].list.includes(name)){ return self.debug('ignoring blacklist', name); }
+			var found = false;
+			_.each(self.configBlacklist[whiteListCheck].patterns, function(pattern){
+				if(!found && pattern.test(name)){
+					found = true;
+				}
+			});
+			if(found){ return self.debug('ignoring blacklist pattern', name); }
+			self.log('Setting %s ENV variable [%s]', whiteListCheck, name);
+		}
+		if(name && !config){
+			data = name;
+		}else{
+			data[name] = config;
+		}
+		switch(whiteListCheck){
+			case "env":
+			case "secrets":
+			case "config-hoist":
+				return _.set(self.config, name, config);
+			default:
+				return _.merge(self.config, _.cloneDeep(data));
+		}
+	}
+
+	loadConfig(cwd, ignoreExtra){
+		var self = this;
+		cwd = cwd || self.cwd;
+		ignoreExtra = ignoreExtra || false;
+
+		// load plugin defaults
+		_.each(self.plugins, function(plugin){
+			if(plugin.config){
+				// ensure sideloaded plugins retain original config
+				if(plugin.original_namespace){
+					plugin.config[plugin.namespace] = plugin.config[plugin.original_namespace];
+					delete plugin.config[plugin.original_namespace];
+				}
+				self.registerConfig(plugin.config);
+			}
+		});
+
+		// load local json files
+		_.each(self.recursiveList(cwd + '/config', '.json'), function(file){
+			// prevent loading base config and codes
+			if(file.indexOf(self.configFile) === -1 && file.indexOf(self.config.codes) === -1){
+				if(!self.config[path.parse(file).name]){
+					self.config[path.parse(file).name] = {};
+				}
+				self.registerConfig(path.parse(file).name, require(file));
+			}
+		});
+
+		if(!ignoreExtra){
+			// handle process flags
+			self.args = minimist(process.argv.slice(2));
+			_.each(self.args, function(value, key){
+				return self.registerConfig(key, value, 'args');
+			});
+			self.argv = _.clone(self.args._) || [];
+
+			// handle environment variables
+			_.each(process.env, function(value, key){
+				key = key.replace(/__/g, '.'); // replace double underscores to dots, to allow object notation in environment vars
+				try{
+					value = JSON.parse(value);
+				}catch(e){
+					// do nothing
+				}
+				return self.registerConfig(key, value, 'env');
+			});
+
+			if(self.config.secrets){
+				// handle docker secrets
+				_.each(self.recursiveList(self.config.secrets, false), function(file){
+					var key, value;
+					try{
+						key = path.basename(file);
+						value = fs.readFileSync(file, 'utf8');
+						value = JSON.parse(value); // if it fails it will revert to above value
+					}catch(e){
+						// do nothing
+					}
+					if(!value || !key){ return; }
+					return self.registerConfig(key, value, 'secrets');
+				});
+			}
+		}else{
+			self.debug("Ignoring config extra loading");
+		}
+		self.emit('app.setup.loadConfig');
+
+		if(self.config.configOverride){
+			// allow dev-config.json in root directory to override config vars
+			var access = null;
+			try{
+				access = require(path.join(self.cwd, self.config.configOverride));
+			}catch(err){
+				// do nothing
+			}
+			if(access){
+				self.debug('Overriding config with custom overrides');
+				_.each(access, function(value, key){
+					return self.registerConfig(key, value, 'config-hoist');
+				});
+			}
+		}
+		return this;
+	}
+
+	initCodes(){
+		var self = this;
+		self.codes = {};
+		_.each(self.recursiveList(__dirname + '/codes', '.json'), function(file){
+			_.merge(self.codes, require(file));
+		});
+		self.emit('app.setup.initCodes');
+		return this;
+	}
+
+	loadCodes(cwd){
+		var self = this;
+		cwd = cwd || self.cwd + self.config.codes;
+
+		// load plugin defaults
+		_.each(self.plugins, function(plugin){
+			if(plugin.codes){
+				self.registerCodes(plugin.codes);
+			}
+		});
+
+		// handle local files
+		var list = null;
+		try{
+			list = self.recursiveList(cwd, ['.json']);
+		}catch(err){
+			self.debug('No codes folder found (%s), skipping', self.config.codes);
+		}
+		if(list){
+			_.each(list, function(file){
+				self.registerCodes(require(file));
+			});
+		}
+		self.emit('app.setup.loadCodes');
+		return this;
+	}
+
+	registerCodes(codes){
+		var self = this;
+		_.merge(self.codes, _.cloneDeep(codes));
+	}
+
+	initRegistry(){
+		var self = this;
+		self.register = [];
+
+		self.on('app.ready', function(){
+			self.status.running = true;
+			// only handle uncaught exceptions when ready
+			process.on('uncaughtException', function(err){
+				self.error(err.message || err).debug(err.stack || '(no stack trace)');
+				// close the app if we have not completed startup
+				if(!self.status.running){
+					self.emit('app.stop', true);
+				}
+			});
+		});
+
+		// app registry is used to track graceful halting
+		self.on('app.register', function(item){
+			if(self.register.indexOf(item) === -1){
+				self.log('Plugin registered: %s', item);
+				self.register.push(item);
+			}
+		});
+		self.on('app.deregister', function(item){
+			var i = self.register.indexOf(item);
+			if(i !== -1){
+				self.register.splice(i, 1);
+				self.warn('De-registered: %s', item);
+			}
+			if(!self.status.running && self.register.length < 1){
+				self.emit('app.exit', true);
+			}
+		});
+		self.on('app.stop', function(){
+			if(self.status.stopping){
+				self.status.stopAttempts++;
+				if(self.status.stopAttempts === 1){
+					self.warn('%s will be closed in %sms if it does not shut down gracefully.', self.config.name, self.config.stopTimeout);
+					setTimeout(function(){
+						self.error('%s took too long to close. Killing process.', self.config.name);
+						self.emit('app.exit');
+					}, self.config.stopTimeout);
+				}
+				if(self.status.stopAttempts < self.config.stopAttempts){
+					return self.warn('%s already stopping. Attempt %s more times to kill process', self.config.name, self.config.stopAttempts - self.status.stopAttempts);
+				}else{
+					self.error('Forcefully killing %s', self.config.name);
+					return self.emit('app.exit');
+				}
+			}
+
+			self.status.running = false;
+			self.status.stopping = true;
+			self.info('Stopping %s gracefully', self.config.name);
+			self.emit('app.close');
+			if(!self.register.length){
+				return self.emit('app.exit', true);
+			}
+		});
+		self.on('app.exit', function(graceful){
+			/* eslint-disable no-process-exit */
+			if(!graceful){
+				return process.exit(1);
+			}
+			self.info('%s gracefully closed.', self.config.name);
+			process.exit();
+		});
+
+		// gracefully handle ctrl+c
+		_.each(['SIGINT', 'SIGUSR2'], function(event){
+			process.on(event, function(){
+				self.emit('app.stop');
+			});
+		});
+
+		// set debug mode on SIGUSR1
+		process.on('SIGUSR1', function(){
+			self.config.debug = !self.config.debug;
+		});
+
+		self.emit('app.setup.initRegistry');
+		return this;
+	}
+
+	initLimitListeners(){
+		var self = this;
+		if(!this.config.trackErrors){ return this; }
+		var issues = {
+			errorCode: {},
+			failCode: {}
+		};
+		_.each(['errorCode', 'failCode'], function(type){
+			function limitToErrors(error){
+				if(!self.limitMaps[type] || !self.limitMaps[type][error.code]){
+					return; // no issue being tracked
+				}
+
+				var limit = self.limitMaps[type][error.code];
+
+				// new issue
+				if(!issues[type][error.code]){
+					issues[type][error.code] = {
+						occurrences: 0, // long count, track
+						balance: 0, // time-based balance
+						dateFirst: moment().unix(),
+						dateLast: null,
+						datesTriggered: [],
+						triggered: false // track if we've triggered the current balance
+					};
+				}
+				issues[type][error.code].occurrences++;
+				issues[type][error.code].balance++;
+				issues[type][error.code].dateLast = moment().unix();
+
+				if(limit.time){
+					setTimeout(function(){
+						issues[type][error.code].balance--;
+						if(issues[type][error.code].balance <= 0){
+							issues[type][error.code].balance = 0;
+							issues[type][error.code].triggered = false;
+						}
+					}, limit.time);
+				}
+				if(!issues[type][error.code].triggered && issues[type][error.code].balance > limit.threshold){
+					issues[type][error.code].triggered = true;
+					limit.callback(_.clone(issues[type][error.code]));
+					issues[type][error.code].datesTriggered.push(issues[type][error.code].dateLast); // add after callback, to avoid double dates
+				}
+			}
+			self.on(type, limitToErrors);
+		});
+		self.emit('app.setup.initLimitListeners');
+		return this;
+	}
+
+	loadPlugins(){
+		var self = this;
+		self.config.plugins = _.map(self.config.plugins, function(plugin){
+			if(typeof(plugin) === 'string'){
+				plugin = {
+					plugin: plugin,
+					name: null,
+					namespace: null
+				};
+			}
+			var pluginFile = require(plugin.plugin);
+			// remove node modules cache to allow reuse of plugins under new namespaces
+			delete require.cache[require.resolve(plugin.plugin)];
+
+			if(plugin.namespace){
+				pluginFile.original_namespace = pluginFile.namespace;
+				plugin.original_namespace = pluginFile.namespace;
+				pluginFile.namespace = plugin.namespace;
+				pluginFile.name = plugin.name;
+				self.info('Sideloading [%s] as plugin: [%s]', plugin.namespace, plugin.name);
+			}else{
+				plugin.namespace = pluginFile.namespace;
+				plugin.name = pluginFile.name;
+				self.info('Loading plugin: [%s]', plugin.name);
+			}
+			self.plugins[plugin.namespace] = pluginFile;
+			return plugin;
+		});
+		self.emit('app.setup.loadPlugins');
+		return this;
+	}
+
+	loadErrorMap(){
+		var self = this;
+		_.each(self.plugins, function(plugin){
+			if(plugin.errors){
+				self.registerErrors(plugin.errors);
+			}
+		});
+		self.emit('app.setup.loadErrorMap');
+		return this;
+	}
+
+	registerPlugin(opts){
+		var self = this;
+		if(self.status.setup){
+			throw new self._errorCode('app.register_plugin_on_runtime');
+		}
+		assert(opts.name, 'Plugin is missing required `name` option.');
+		assert(opts.namespace, 'Plugin is missing required `namespace` option.');
+		assert(opts.exports, 'Plugin is missing required `exports` function.');
+		//self.logs.prefix = opts.namespace;
+		self.loadConfig(opts.dir, true);
+		self.loadCodes(opts.dir + '/codes');
+
+		return _.merge(opts, {
+			codes: self.codes || null,
+			config: self.config || null
+		});
+	}
+
+	setupJSONHandler(){
+		require(__dirname + '/require-extensions.js');
+		return this;
+	}
+
+	registerLimit(code, threshold, options, callback){
+		if(!callback && options){
+			callback = options;
+			options = {};
+		}
+		var opts = _.defaults(options, {
+			callback: callback,
+			threshold: threshold,
+			error: 'errorCode', // or failCode
+			index: null, // 'object.to.path' of unique index to track by
+			reset: true, // reset balance counter on callback
+			time: null
+		});
+
+		if(!this.limitMaps[opts.error]){
+			this.limitMaps[opts.error] = {};
+		}
+		if(!this.limitMaps[opts.error][code]){
+			this.limitMaps[opts.error][code] = [];
+		}
+		this.limitMaps[opts.error][code].push(opts);
+	}
+}
+
+module.exports = spawnpoint;
